requirements.txt
python-telegram-bot[job-queue]
python-dotenv
requests
caldav
vobject
google-generativeai
